fix: componentで置き換え。一覧ページからチケットを取得できないバグが存在
